fix more dml warnings (#21980)
### Description
Fixes more warnings in DML execution provider that lead to security
issues in binskim


### Motivation and Context
OS components that include ORT must treat certain warnings as errors,
and cannot disable critical compiler warnings

https://github.com/microsoft/binskim/blob/main/src/BinSkim.Rules/PERules/BA2007.EnableCriticalCompilerWarnings.cs

+++ onnxruntime/core/providers/dml/DmlExecutionProvider/src/DmlGraphFusionHelper.cpp
@@ -368,7 +368,7 @@ namespace DmlGraphFusionHelper
 
                     DML_CONSTANT_DATA_GRAPH_NODE_DESC* constantNode = allocator.template Allocate<DML_CONSTANT_DATA_GRAPH_NODE_DESC>();
                     constantNode->Name = node.Name.data();
-                    constantNode->DataSize = constantData.dataSize;
+                    constantNode->DataSize = gsl::narrow_cast<size_t>(constantData.dataSize);
                     constantNode->Data = constantData.data;
                     dmlGraphNodes.push_back(DML_GRAPH_NODE_DESC{DML_GRAPH_NODE_TYPE_CONSTANT, constantNode});
                 }
@@ -845,11 +845,11 @@ namespace DmlGraphFusionHelper
             .Provider(onnxruntime::kDmlExecutionProvider);
 
         // Force the CPU inputs to be allocated on the CPU
-        for (int i = 0; i < subGraphInputArgNames.size(); ++i)
+        for (size_t i = 0; i < subGraphInputArgNames.size(); ++i)
         {
             if (dynamicCpuInputMap.find(subGraphInputArgNames[i]) != dynamicCpuInputMap.end())
             {
-                builder.InputMemoryType(OrtMemTypeCPUInput, i);
+                builder.InputMemoryType(OrtMemTypeCPUInput, static_cast<int>(i));
             }
         }
